Add request time and full dependency list
Not all of the dependencies were revealed that needed to be installed at
the top of the stats file. These have now been added. I also added the
request time for the items that were already added as it makes it easier
to follow in a long run where you are at as well as diagnosing issues.